+++ cursome/src/main/java/br/com/luciano/cursome/services/CategoriaService.java
@@ -20,7 +20,7 @@ public class CategoriaService {
     @Autowired
     private CategoriaRepository categoriaRepository;
 
-    public Categoria buscar(long id) {
+    public Categoria buscar(Integer id) {
         Optional<Categoria> obj = categoriaRepository.findById(id); 
         return obj.orElse(null); 
     }
